Refactor logging package: * move store implementation to logging.sqlalchemy, remove dependency to SQLiteConfig * move SQLiteConfig to logging.logger * remove obsolete file logging.sqlite * rename test files to mirror subpackage structure
