+++ tests/optimagic/logging/test_logger.py
